Add simple readme
I based the format on this example from dbader: https://github.com/dbader/readme-template/blob/master/README.md. We should update this with additional information in the future.

For the record, I felt a little guilty including my name on this document, but I do plan on contributing more to the project in the future. 🤥