Add IconSelector to CauseForm
Icons scroll from left to right as opposited to using navigation buttons
to preserve the dimensions of the icons.

The meat of the IconSelector can probably be converted into separate
subcomponents, but it functions for now.

Additionally, I added some more variables to index.css and imported a
new font. Raleway more closely matches the design.